* master:
  typos
  udpated notification doc
  typos
  updated docs
  Casting age and value
  Integer casts
  Made NSInteger casts
  Casted macro on testVersionGetters
  Removed duplicate RLMRealm accessor method from RLMTestCase

Conflicts:
	src/realm/objc/test/group.m
	src/realm/objc/test/group_misc_2.m
	src/realm/objc/test/tutorial.m

+++ src/realm/objc/test/dynamic_table.mm
@@ -1525,8 +1525,8 @@ using namespace std;
         
         
         // Test int min
-        XCTAssertEqual([[table minOfProperty:@"IntCol" where:@"BoolCol == NO"] integerValue], 2, @"Minimum should be 2");
-        XCTAssertEqual([[table minOfProperty:@"IntCol" where:@"BoolCol == YES"] integerValue], 1, @"Minimum should be 1");
+        XCTAssertEqual([[table minOfProperty:@"IntCol" where:@"BoolCol == NO"] integerValue], (NSInteger)2, @"Minimum should be 2");
+        XCTAssertEqual([[table minOfProperty:@"IntCol" where:@"BoolCol == YES"] integerValue], (NSInteger)1, @"Minimum should be 1");
         
         // Test float min
         XCTAssertEqualWithAccuracy([[table minOfProperty:@"FloatCol" where:@"BoolCol == NO"] floatValue], (float)1.2f, 0.1f, @"Minimum should be 1.2f");
@@ -1543,8 +1543,8 @@ using namespace std;
         XCTAssertThrows([table minOfProperty:@"BoolCol" where:@"IntCol == 1"], @"Should throw exception");
         
         // Test int max
-        XCTAssertEqual([[table maxOfProperty:@"IntCol" where:@"BoolCol == NO"] integerValue], 8, @"Maximum should be 8");
-        XCTAssertEqual([[table maxOfProperty:@"IntCol" where:@"BoolCol == YES"] integerValue], 10, @"Maximum should be 10");
+        XCTAssertEqual([[table maxOfProperty:@"IntCol" where:@"BoolCol == NO"] integerValue], (NSInteger)8, @"Maximum should be 8");
+        XCTAssertEqual([[table maxOfProperty:@"IntCol" where:@"BoolCol == YES"] integerValue], (NSInteger)10, @"Maximum should be 10");
         
         // Test float max
         XCTAssertEqualWithAccuracy([[table maxOfProperty:@"FloatCol" where:@"BoolCol == NO"] floatValue], (float)1.8f, 0.1f, @"Maximum should be 1.8f");
